pyproject.toml
[tool.poetry]
name = "scGPT"
version = "0.2.1"
description = "Large-scale generative pretrain of single cell using transformer."
license = "MIT"
authors = ["Haotian <subercui@gmail.com>"]
repository = "https://github.com/bowang-lab/scGPT"
homepage = "https://github.com/bowang-lab/scGPT"
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.10,<3.11" 
pandas = ">=1.3.5"
scvi-tools = ">=0.16.0,<1.0"
scanpy = "^1.9.1"
torch = ">=1.13.0"
torchtext = "*"
numba = ">=0.55.1"
scikit-misc = ">=0.1.4"
umap-learn = ">=0.5.3"
leidenalg = ">=0.8.10"
datasets = "^2.3.0"
typing-extensions = "^4.2.0"
scib = "^1.0.3"
cell-gears = "<0.0.3"
orbax = "<0.1.8"
jaxlib = "0.4.13"
scipy = "^1.14.1"
scikit-learn = "^1.5.2"
tqdm = "^4.66.6"

[tool.poetry.dev-dependencies]
pytest = "^5.2"
black = "^22.3.0"
tensorflow = "^2.8.0"
torch-geometric = "^2.3.0"
dcor = "~0.5.3"
wandb = ">=0.18"
plotly = "^5.3.1"


[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
